% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Darwin.R
\name{darwinize}
\alias{darwinize}
\title{Darwin's naturalization conundrum}
\usage{
darwinize(x, phy, iter)
}
\arguments{
\item{x}{A data frame with list of species}

\item{phy}{A phylogenetic tree}

\item{iter}{Numeric, the number of permutations}
}
\value{
a vector with the mean distance between species
}
\description{
\code{darwinize} computes standard effective size of
distance to nearest neighbors.
}
\author{
Barnabas H. Daru \email{darunabas@gmail.com}
}
\keyword{bioregion}
