% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_climate.R
\name{extract_climate}
\alias{extract_climate}
\title{Extracts environmental variables}
\usage{
extract_climate(x, y, val = "val", FUN = sd, ...)
}
\arguments{
\item{x}{A raster file of the environmental variable}

\item{y}{Spatial polygons data frame corresponding to
the study area of interest to extract the variable.}

\item{val}{String to rename the column with the extracted
variable.}

\item{FUN}{The function to summarize the values (e.g. sd, mean)}

\item{\dots}{arguments passed among methods.}
}
\value{
A list or data frame.
}
\description{
This function extracts environmental variables from raster
files within spatial spatial polygons.
}
\author{
Barnabas H. Daru \email{darunabas@gmail.com}
}
