% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_names.R
\name{family_name}
\alias{family_name}
\title{Get APG family names of plants}
\usage{
family_name(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
a vector with the family names
}
\description{
This function generates family names based on
the Angiosperm Phylogeny Group system.
}
\keyword{bioregion}
