% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd2grids.R
\name{odd2grids}
\alias{odd2grids}
\title{Convert irregular community data to grids}
\usage{
odd2grids(x, shp, res = 0.25)
}
\arguments{
\item{x}{input data framme.}

\item{shp}{the irregular polygon shapefile to be converted}

\item{res}{the grain size of the grid cells in decimal degrees (default).}
}
\value{
\itemize{
  \item comm_dat: community data frame
  \item poly_shp: shapefile of grid cells with the values per cell.
}
}
\description{
This functions converts irregular polygon shapefiles to grids
}
