% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Darwin.R
\name{darwin_hypothesis}
\alias{darwin_hypothesis}
\title{Darwin's naturalization conundrum}
\usage{
darwin_hypothesis(dat, phy, iter, grids = "grids", species = "species")
}
\arguments{
\item{dat}{A data frame}

\item{phy}{A phylogenetic tree}

\item{iter}{Numeric, the number of permutations}

\item{grids}{The column with the grid cells}

\item{species}{The column with the species names}
}
\value{
a vector with the mean distance between species
}
\description{
\code{darwin_hypothesis} computes standard effective size of
distance to nearest neighbors.
}
\author{
Barnabas H. Daru \email{darunabas@gmail.com}
}
\keyword{bioregion}
